move the redirection to the dedicated class and move the loading into the wpMatomo class to normalize hooks registrations
