+++ classes/WpMatomo/RedirectOnActivation.php
@@ -8,14 +8,17 @@
  */
 
 namespace WpMatomo;
+
+use \WpMatomo\Admin\TrackingSettings;
+
 class RedirectOnActivation {
 	/**
-	 * @var WpMatomo
+	 * @var Settings
 	 */
-	public $wpMatomo;
+	public static $settings;
 
-	public function __construct(\WpMatomo $wpMatomo) {
-		$this->wpMatomo = $wpMatomo;
+	public function __construct() {
+		self::$settings = new Settings();
 	}
 
 	public function register_hooks() {
@@ -31,7 +34,29 @@ class RedirectOnActivation {
 	public function matomo_plugin_redirect() {
 		if ( get_option( 'matomo_plugin_do_activation_redirect', false ) ) {
 			delete_option( 'matomo_plugin_do_activation_redirect' );
-			$this->wpMatomo->redirect_to_getting_started();
+			$this->redirect_to_getting_started();
 		}
+	}
+	/**
+	 * We don't test the result of the wp_redirect method and we silent this method
+	 * as this method will not work during unit tests.
+	 * We just return if yes or no we should redirect
+	 *
+	 * @see https://github.com/matomo-org/matomo-for-wordpress/issues/434
+	 * @return boolean
+	 */
+	public function redirect_to_getting_started() {
+		$redirect = false;
+		if(!isset($_GET['activate-multi'])) {
+			if
+			(
+				( self::$settings->get_global_option( Settings::SHOW_GET_STARTED_PAGE ) === 1 ) &&
+				( self::$settings->get_global_option( 'track_mode' ) === TrackingSettings::TRACK_MODE_DISABLED )
+			) {
+				$redirect = true;
+				@wp_redirect( admin_url( 'admin.php?page=matomo-get-started' ) );
+			}
+		}
+		return $redirect;
 	}
 }
